+++ dune/istl/umfpack.hh
@@ -17,6 +17,7 @@
 #include<dune/istl/bcrsmatrix.hh>
 #include<dune/istl/solvers.hh>
 #include<dune/istl/solvertype.hh>
+#include <dune/istl/solverfactory.hh>
 
 #include"colcompmatrix.hh"
 
@@ -574,7 +575,36 @@ namespace Dune {
   {
     enum { value = true };
   };
-}
+
+  struct UMFPackCreator {
+    template<class F,class=void> struct isValidBlock : std::false_type{};
+    template<class B> struct isValidBlock<B, std::enable_if_t<std::is_same<typename FieldTraits<B>::real_type,double>::value>> : std::true_type {};
+
+    template<typename TL, typename M>
+    std::shared_ptr<Dune::InverseOperator<typename Dune::TypeListElement<1, TL>::type,
+                                          typename Dune::TypeListElement<2, TL>::type>>
+    operator() (TL /*tl*/, const M& mat, const Dune::ParameterTree& config,
+      std::enable_if_t<
+                isValidBlock<typename Dune::TypeListElement<1, TL>::type::block_type>::value,int> = 0) const
+    {
+      int verbose = config.get("verbose", 0);
+      return std::make_shared<Dune::UMFPack<M>>(mat,verbose);
+    };
+
+    // second version with SFINAE to validate the template parameters of UMFPack
+    template<typename TL, typename M>
+    std::shared_ptr<Dune::InverseOperator<typename Dune::TypeListElement<1, TL>::type,
+                                          typename Dune::TypeListElement<2, TL>::type>>
+    operator() (TL /*tl*/, const M& /*mat*/, const Dune::ParameterTree& /*config*/,
+      std::enable_if_t<
+                !isValidBlock<typename Dune::TypeListElement<1, TL>::type::block_type>::value,int> = 0) const
+    {
+      DUNE_THROW(UnsupportedType,
+        "Unsupported Type in UMFPack (only double and std::complex<double> supported)");
+    };
+  };
+  DUNE_REGISTER_DIRECT_SOLVER("umfpack",Dune::UMFPackCreator());
+} // end namespace Dune
 
 #endif // HAVE_SUITESPARSE_UMFPACK